SONAR-6567 disable extension point NotificationDispatcher in batch
Currently it's only removed from API. Will be removed from batch after that events are in Compute Engine (for change of quality gate status)
