+++ server/sonar-server/src/main/java/org/sonar/server/issue/notification/ChangesOnMyIssueNotificationDispatcher.java
@@ -25,6 +25,9 @@ import org.sonar.api.notifications.*;
 
 import javax.annotation.Nullable;
 import java.util.Collection;
+import org.sonar.core.notification.NotificationDispatcher;
+import org.sonar.core.notification.NotificationDispatcherMetadata;
+import org.sonar.core.notification.NotificationManager;
 
 /**
  * This dispatcher means: "notify me when a change is done on an issue that is assigned to me or reported by me".
