+++ server/sonar-server/src/main/java/org/sonar/server/issue/notification/DoNotFixNotificationDispatcher.java
@@ -27,6 +27,9 @@ import org.sonar.api.notifications.*;
 
 import java.util.Collection;
 import java.util.Map;
+import org.sonar.core.notification.NotificationDispatcher;
+import org.sonar.core.notification.NotificationDispatcherMetadata;
+import org.sonar.core.notification.NotificationManager;
 
 /**
  * This dispatcher means: "notify me when an issue is resolved as false positive or won't fix".
