+++ server/sonar-server/src/main/java/org/sonar/server/issue/notification/NewIssuesNotificationDispatcher.java
@@ -24,6 +24,9 @@ import org.sonar.api.notifications.*;
 
 import java.util.Collection;
 import java.util.Map;
+import org.sonar.core.notification.NotificationDispatcher;
+import org.sonar.core.notification.NotificationDispatcherMetadata;
+import org.sonar.core.notification.NotificationManager;
 
 /**
  * This dispatcher means: "notify me when new issues are introduced during project analysis"
